Custom Tomcat formatter and handler
Previously the output from Tomcat (and other components using JULI)
was very hard to read and printed to stderr.  This resulted in
difficult to read logs, and when printed to Loggergator a scary [ERR]
prefix.  This change adds a custom formatter to output the logs as:

[CONTAINER] c.extra.super.really.long.logger INFO    Test Message

and a custom handler to print that (and all messages) to stdout.

[#64764934]