Split Tomcat support
Previously, all Tomcat buildpack support was in a single package.  Due
to the way that the logging support must be loaded into classloaders,
this would not work.  This change splits the existing support into two
different projects and therefore two different JARs.

Since this is a compatibility-breaking change, the version number of
the project has been revved to 2.0.0.BUILD-SNAPSHOT.

[#64764934]