Add support for a Tomcat Access Logging Valve
The builtin Access Logging Valve in Tomcat will not log to standard
out. This is required in Cloud Foundry in order for logging to be
captured and made easily available to end users. This commit adds a
new support component that provides a Tomcat Access Logging Valve that
logs to standard out.

[#72881930]