Tomcat swallows System.err output
When Tomcat starts up we set the ConsoleHandlers output stream to System.out in
our custom sub-class. The default is System.err and the ConsoleHandler will call
close on it when replacing it with System.out. This commit ensures that the
ConsoleHandler does not call close on the existing output stream by reflectively
nulling it first.

[#80445246]