Java 11 Warnings
Previously a work around was put in place to deal with the fact that setting a
ConsoleHandler's stream after it was configured would close stderr for the
process.  This workaround involved using reflection to solve the problem.  As
Java 9+ has come into usage, the restrictions on reflection have caused
warnings from this behavior.  This change updates the work around, to simply
extend a higher base-class and replicate the ConsoleHandler's behavior.

[resolves #8]